requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.10.5
colorama==0.4.6
distro==1.9.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.11.1
openai==2.6.0
pyaes==1.6.1
pydantic==2.12.3
pydantic_core==2.41.4
Pyrogram==2.0.106
PySocks==1.7.1
python-dotenv==1.1.1
sniffio==1.3.1
TgCrypto==1.2.5
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2